conf/rc.sun: Define PIDFILE not in /var/run (or /run)
If /run or /var/run is on a tmpfs, the default permissions do not
allow for non-root daemons to create a pidfile there (the expectation
is that unprivileged daemons are started by root and then drop privs,
but this is not the case for SUN, nor should it be).

Instead, let's use a location in $HOME - I chose $HOME/.cache/sun/
but that is of course up for debate/discussion. This commit also
ensures the PIDFILE directory is available before attempting to write
the pidfile there.